feat(docker,services): dockerize core services, add healthchecks; fix(AuctionDto): change monetary fields to decimal and update mapping
Summary
- Full Dockerization of services:
  - Auction Service: updated docker-compose with dedicated container.
  - Search Service: added Dockerfile and service entry to compose.
  - Identity Service: containerized with secure environment variable setup.
  - Gateway Service: container with reverse proxy and dependencies.
- Observability and reliability:
  - Healthchecks added in `docker-compose.yml` (start_period, interval, retries).
  - Exposed health endpoints for services (.NET HealthChecks).
- Developer productivity:
  - Enabled debugging for .NET services running in containers (ports, volumes, `ASPNETCORE_ENVIRONMENT=Docker`).
  - Split build targets per service with optimized multi-stage builds (restore/build/publish).
- Domain/DTO quality:
  - **Fix**: changed monetary properties `SoldAmount`, `ReservePrice`, and `CurrentHighBid` in **AuctionDto** from **int** to **decimal**.
  - Updated mapping between **Auction** entity and **AuctionDto** (AutoMapper/Profile) to ensure financial precision and contract consistency.

Key Details
- `docker-compose.yml`:
  - Services: `auction`, `search`, `identity`, `gateway`.
  - Configured `depends_on` and `healthcheck` for robust orchestration.
  - Environment variables and bind mounts configured for debugging (`/app`, `ASPNETCORE_URLS`, `DOTNET_USE_POLLING_FILE_WATCHER`).
- Dockerfiles:
  - Multi-stage builds (sdk → runtime) with restore caching.
  - Standardized `ENTRYPOINT` and correct `EXPOSE` for each service.

Domain/DTO Fixes
- `AuctionDto.SoldAmount: int → decimal`
- `AuctionDto.ReservePrice: int → decimal`
- `AuctionDto.CurrentHighBid: int → decimal`
- AutoMapper/Profile:
  - Updated `Auction ↔ AuctionDto` mappings to `decimal`.
  - Removed explicit conversions; ensured financial precision.

Rationale
- Monetary values require `decimal` to avoid precision loss and rounding errors in bids and auction settlements.
- Healthchecks and standardized containerization improve MTTR and reliability in local/CI environments.

+++ src/AuctionService/Data/Migrations/AuctionDbContextModelSnapshot.cs
@@ -34,17 +34,17 @@ namespace AuctionService.Data.Migrations
                     b.Property<DateTime>("CreatedAt")
                         .HasColumnType("timestamp with time zone");
 
-                    b.Property<int?>("CurrentHighBid")
-                        .HasColumnType("integer");
+                    b.Property<decimal?>("CurrentHighBid")
+                        .HasColumnType("numeric");
 
-                    b.Property<int>("ReservePrice")
-                        .HasColumnType("integer");
+                    b.Property<decimal>("ReservePrice")
+                        .HasColumnType("numeric");
 
                     b.Property<string>("Seller")
                         .HasColumnType("text");
 
-                    b.Property<int?>("SoldAmount")
-                        .HasColumnType("integer");
+                    b.Property<decimal?>("SoldAmount")
+                        .HasColumnType("numeric");
 
                     b.Property<int>("Status")
                         .HasColumnType("integer");
